backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
openai==2.24.0
pytest==8.0.0
pytest-asyncio==0.23.3
PyPDF2==3.0.1
pdfplumber==0.10.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6